Overhaul annual fees and include maintenance fee

+++ GECO/_Code/DecimalFunctions.vb
@@ -3,7 +3,7 @@
 Public Module DecimalFunctions
 
     Public Function NullableDecimalProduct(a As Object, b As Object) As Decimal?
-        Dim aNull As Decimal? = GetNullable(Of Decimal?)(a)
+        Dim aNull As Integer? = GetNullable(Of Integer?)(a)
         Dim bNull As Decimal? = GetNullable(Of Decimal?)(b)
 
         If aNull.HasValue AndAlso bNull.HasValue Then
